useSequencerStateMachineを変更、関連ファイルを整理
