+++ src/sing/sequencerStateMachine/common.ts
