+++ src/sing/sequencerStateMachine/states/resizeNoteLeftState.ts
@@ -1,84 +1,104 @@
-import { SetNextState, State } from "@/sing/stateMachine";
+import { getOrThrow } from "@/helpers/mapHelper";
+import { State, SetNextState } from "@/sing/stateMachine";
+import { getButton, PREVIEW_SOUND_DURATION } from "@/sing/viewHelper";
+import { Note } from "@/store/type";
+import { TrackId, NoteId } from "@/type/preload";
 import {
   Context,
   getGuideLineTicks,
   Input,
   PositionOnSequencer,
   SequencerStateDefinitions,
-} from "@/composables/sequencerStateMachine/common";
-import { NoteId, TrackId } from "@/type/preload";
-import { Note } from "@/store/type";
-import {
-  getButton,
-  getDoremiFromNoteNumber,
-  PREVIEW_SOUND_DURATION,
-} from "@/sing/viewHelper";
+} from "@/sing/sequencerStateMachine/common";
 import { clamp } from "@/sing/utility";
 
-export class AddNoteState
+export class ResizeNoteLeftState
   implements State<SequencerStateDefinitions, Input, Context>
 {
-  readonly id = "addNote";
+  readonly id = "resizeNoteLeft";
 
   private readonly cursorPosAtStart: PositionOnSequencer;
   private readonly targetTrackId: TrackId;
+  private readonly targetNoteIds: Set<NoteId>;
+  private readonly mouseDownNoteId: NoteId;
 
   private currentCursorPos: PositionOnSequencer;
+
   private innerContext:
     | {
-        noteToAdd: Note;
+        targetNotesAtStart: Map<NoteId, Note>;
         previewRequestId: number;
         executePreviewProcess: boolean;
+        edited: boolean;
+        guideLineTicksAtStart: number;
       }
     | undefined;
 
   constructor(args: {
     cursorPosAtStart: PositionOnSequencer;
     targetTrackId: TrackId;
+    targetNoteIds: Set<NoteId>;
+    mouseDownNoteId: NoteId;
   }) {
+    if (!args.targetNoteIds.has(args.mouseDownNoteId)) {
+      throw new Error("mouseDownNoteId is not included in targetNoteIds.");
+    }
     this.cursorPosAtStart = args.cursorPosAtStart;
     this.targetTrackId = args.targetTrackId;
+    this.targetNoteIds = args.targetNoteIds;
+    this.mouseDownNoteId = args.mouseDownNoteId;
 
     this.currentCursorPos = args.cursorPosAtStart;
   }
 
-  private previewAdd(context: Context) {
+  private previewResizeLeft(context: Context) {
     if (this.innerContext == undefined) {
       throw new Error("innerContext is undefined.");
     }
-    const noteToAdd = this.innerContext.noteToAdd;
     const snapTicks = context.snapTicks.value;
-    const dragTicks = this.currentCursorPos.ticks - this.cursorPosAtStart.ticks;
-    const noteDuration = Math.round(dragTicks / snapTicks) * snapTicks;
-    const noteEndPos = noteToAdd.position + noteDuration;
     const previewNotes = context.previewNotes.value;
+    const targetNotesAtStart = this.innerContext.targetNotesAtStart;
+    const mouseDownNote = getOrThrow(targetNotesAtStart, this.mouseDownNoteId);
+    const dragTicks = this.currentCursorPos.ticks - this.cursorPosAtStart.ticks;
+    const notePos = mouseDownNote.position;
+    const newNotePos =
+      Math.round((notePos + dragTicks) / snapTicks) * snapTicks;
+    const movingTicks = newNotePos - notePos;
 
     const editedNotes = new Map<NoteId, Note>();
     for (const note of previewNotes) {
-      const duration = Math.max(snapTicks, noteDuration);
-      if (note.duration !== duration) {
-        editedNotes.set(note.id, { ...note, duration });
+      const targetNoteAtStart = getOrThrow(targetNotesAtStart, note.id);
+      const notePos = targetNoteAtStart.position;
+      const noteEndPos =
+        targetNoteAtStart.position + targetNoteAtStart.duration;
+      const position = clamp(notePos + movingTicks, 0, noteEndPos - snapTicks);
+      const duration = noteEndPos - position;
+
+      if (note.position !== position || note.duration !== duration) {
+        editedNotes.set(note.id, { ...note, position, duration });
       }
     }
     if (editedNotes.size !== 0) {
       context.previewNotes.value = previewNotes.map((value) => {
         return editedNotes.get(value.id) ?? value;
       });
+      this.innerContext.edited = true;
     }
-    context.guideLineTicks.value = noteEndPos;
+
+    context.guideLineTicks.value = newNotePos;
   }
 
   onEnter(context: Context) {
     const guideLineTicks = getGuideLineTicks(this.cursorPosAtStart, context);
-    const noteToAdd = {
-      id: NoteId(crypto.randomUUID()),
-      position: Math.max(0, guideLineTicks),
-      duration: context.snapTicks.value,
-      noteNumber: clamp(this.cursorPosAtStart.noteNumber, 0, 127),
-      lyric: getDoremiFromNoteNumber(this.cursorPosAtStart.noteNumber),
-    };
+    const targetNotesArray = context.notesInSelectedTrack.value.filter(
+      (value) => this.targetNoteIds.has(value.id),
+    );
+    const targetNotesMap = new Map<NoteId, Note>();
+    for (const targetNote of targetNotesArray) {
+      targetNotesMap.set(targetNote.id, targetNote);
+    }
 
-    context.previewNotes.value = [noteToAdd];
+    context.previewNotes.value = [...targetNotesArray];
     context.nowPreviewing.value = true;
 
     const previewIfNeeded = () => {
@@ -86,7 +106,7 @@ export class AddNoteState
         throw new Error("innerContext is undefined.");
       }
       if (this.innerContext.executePreviewProcess) {
-        this.previewAdd(context);
+        this.previewResizeLeft(context);
         this.innerContext.executePreviewProcess = false;
       }
       this.innerContext.previewRequestId =
@@ -95,9 +115,11 @@ export class AddNoteState
     const previewRequestId = requestAnimationFrame(previewIfNeeded);
 
     this.innerContext = {
-      noteToAdd,
+      targetNotesAtStart: targetNotesMap,
       executePreviewProcess: false,
       previewRequestId,
+      edited: false,
+      guideLineTicksAtStart: guideLineTicks,
     };
   }
 
@@ -117,10 +139,11 @@ export class AddNoteState
       if (input.mouseEvent.type === "mousemove") {
         this.currentCursorPos = input.cursorPos;
         this.innerContext.executePreviewProcess = true;
-      } else if (input.mouseEvent.type === "mouseup") {
-        if (mouseButton === "LEFT_BUTTON") {
-          setNextState("idle", undefined);
-        }
+      } else if (
+        input.mouseEvent.type === "mouseup" &&
+        mouseButton === "LEFT_BUTTON"
+      ) {
+        setNextState("idle", undefined);
       }
     }
   }
@@ -134,8 +157,8 @@ export class AddNoteState
 
     cancelAnimationFrame(this.innerContext.previewRequestId);
 
-    void context.store.actions.COMMAND_ADD_NOTES({
-      notes: context.previewNotes.value,
+    void context.store.actions.COMMAND_UPDATE_NOTES({
+      notes: previewNotes,
       trackId: this.targetTrackId,
     });
     void context.store.actions.SELECT_NOTES({ noteIds: previewNoteIds });
